Developed registrationPage, create new components
Developed registartion form and registration page. Create inputPassword component, formLinks component. Refactor forms.
