+++ src/components/ui/inputPassword/inputPassword.module.scss
@@ -1,0 +1,29 @@
+@import "assets/styles/constants.scss";
+
+.label {
+    color: $gray;
+    opacity: .8;
+    margin: 10px 0;
+}
+
+.passwordInputWrapper {
+    position: relative;
+}
+
+.passwordSvgWrapper {
+    position: absolute;
+    top: 5px;
+    right: 7px;
+    cursor: pointer;
+}
+
+.input_password {
+    padding: 10px 40px 10px 10px;
+    width: calc(100% - 50px);
+    border-radius: 10px;
+    border: none;
+    outline: none;
+    margin-bottom: 20px;
+    background-color: $dark-gray;
+    color: #000000;
+}
