+++ src/pages/auth/formPage/LoginPage.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
-import { NavLink } from 'react-router-dom';
-import { paths } from 'routing/config';
 
-import { LoginForm } from 'components/pages/auth/loginForm';
+import { FormLinks } from 'components/pages/auth/FormLinks';
+import { LoginForm } from 'components/pages/auth/LoginForm';
 
 import styles from 'pages/auth/formPage/auth.module.scss'
 
@@ -10,28 +9,7 @@ export const LoginPage: React.FC = () => {
 
     return (
         <section className = {styles.sectionForm}>
-            <div className = {styles.linksWrapper}>
-                <NavLink 
-                    className={({ isActive }) => {
-                        const linkClasses = [styles.links];
-                        if (isActive) linkClasses.push(styles.active);
-                        return linkClasses.join(" "); 
-                    }} 
-                    to = {paths.loginPage}
-                    >
-                    SIGN IN
-                </NavLink>
-                <NavLink 
-                    className={({ isActive }) => {
-                        const linkClasses = [styles.links];
-                        if (isActive) linkClasses.push(styles.active);
-                        return linkClasses.join(" "); 
-                    }} 
-                    to = {paths.registrationPage}
-                    >
-                    SIGN UP
-                </NavLink>
-            </div>
+            <FormLinks />
             <LoginForm />
         </section>
     );
