+++ src/pages/auth/formPage/auth.module.scss
@@ -8,24 +8,3 @@
     box-shadow: 0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
     border-radius: 10px;
 }
-
-.linksWrapper {
-    margin: 20px 0 40px;
-    :first-child {
-        margin-right: 30px;
-    }
-}
-
-.links {
-    color: $green;
-    opacity: .5;
-    padding-bottom: 3px;
-    font-size: 20px;
-}
-
-.active {
-    color: $green;
-    opacity: 1;
-    border-bottom: 4px solid $green;
-    border-radius: 4px;
-}
